fix: update chunk index types from UInt8 to UInt16
Core widened chunk indices from u8 to u16 to support 32KB+ payloads
(400+ chunks at 80 bytes each exceeds u8 max of 255).

+++ Vauchi/Views/FaceToFaceExchangeView.swift
@@ -434,7 +434,7 @@ struct FaceToFaceExchangeView: View {
         }
     }
 
-    private func transferProgressText(received: UInt8, total: UInt8) -> String {
+    private func transferProgressText(received: UInt16, total: UInt16) -> String {
         if total > 0 {
             return "Receiving \(received)/\(total) chunks..."
         }
